# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Web3 Global Integration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChainStatus = (typeof ChainStatus)[keyof typeof ChainStatus];

export const ChainStatus = {
  active: "active",
  inactive: "inactive",
  testing: "testing",
} as const;

export interface Chain {
  id: number;
  name: string;
  symbol: string;
  chainId?: number;
  rpcUrl?: string;
  status: ChainStatus;
  tvl?: string;
  agentCount?: number;
  createdAt: string;
}

export type CreateChainBodyStatus =
  (typeof CreateChainBodyStatus)[keyof typeof CreateChainBodyStatus];

export const CreateChainBodyStatus = {
  active: "active",
  inactive: "inactive",
  testing: "testing",
} as const;

export interface CreateChainBody {
  name: string;
  symbol: string;
  chainId?: number;
  rpcUrl?: string;
  status: CreateChainBodyStatus;
  tvl?: string;
}

export type PartnershipType =
  (typeof PartnershipType)[keyof typeof PartnershipType];

export const PartnershipType = {
  protocol: "protocol",
  infrastructure: "infrastructure",
  dao: "dao",
  exchange: "exchange",
  launchpad: "launchpad",
  other: "other",
} as const;

export type PartnershipStatus =
  (typeof PartnershipStatus)[keyof typeof PartnershipStatus];

export const PartnershipStatus = {
  active: "active",
  pending: "pending",
  negotiating: "negotiating",
  closed: "closed",
} as const;

export interface Partnership {
  id: number;
  name: string;
  type: PartnershipType;
  status: PartnershipStatus;
  description?: string;
  website?: string;
  logoUrl?: string;
  chainIds?: number[];
  createdAt: string;
}

export type CreatePartnershipBodyType =
  (typeof CreatePartnershipBodyType)[keyof typeof CreatePartnershipBodyType];

export const CreatePartnershipBodyType = {
  protocol: "protocol",
  infrastructure: "infrastructure",
  dao: "dao",
  exchange: "exchange",
  launchpad: "launchpad",
  other: "other",
} as const;

export type CreatePartnershipBodyStatus =
  (typeof CreatePartnershipBodyStatus)[keyof typeof CreatePartnershipBodyStatus];

export const CreatePartnershipBodyStatus = {
  active: "active",
  pending: "pending",
  negotiating: "negotiating",
  closed: "closed",
} as const;

export interface CreatePartnershipBody {
  name: string;
  type: CreatePartnershipBodyType;
  status: CreatePartnershipBodyStatus;
  description?: string;
  website?: string;
  logoUrl?: string;
  chainIds?: number[];
}

export type EcosystemProjectCategory =
  (typeof EcosystemProjectCategory)[keyof typeof EcosystemProjectCategory];

export const EcosystemProjectCategory = {
  defi: "defi",
  nft: "nft",
  gaming: "gaming",
  dao: "dao",
  infrastructure: "infrastructure",
  bridge: "bridge",
  oracle: "oracle",
  other: "other",
} as const;

export type EcosystemProjectStatus =
  (typeof EcosystemProjectStatus)[keyof typeof EcosystemProjectStatus];

export const EcosystemProjectStatus = {
  live: "live",
  building: "building",
  planned: "planned",
  paused: "paused",
} as const;

export interface EcosystemProject {
  id: number;
  name: string;
  category: EcosystemProjectCategory;
  status: EcosystemProjectStatus;
  description?: string;
  website?: string;
  logoUrl?: string;
  chainIds?: number[];
  tvl?: string;
  users?: number;
  createdAt: string;
}

export type CreateEcosystemProjectBodyCategory =
  (typeof CreateEcosystemProjectBodyCategory)[keyof typeof CreateEcosystemProjectBodyCategory];

export const CreateEcosystemProjectBodyCategory = {
  defi: "defi",
  nft: "nft",
  gaming: "gaming",
  dao: "dao",
  infrastructure: "infrastructure",
  bridge: "bridge",
  oracle: "oracle",
  other: "other",
} as const;

export type CreateEcosystemProjectBodyStatus =
  (typeof CreateEcosystemProjectBodyStatus)[keyof typeof CreateEcosystemProjectBodyStatus];

export const CreateEcosystemProjectBodyStatus = {
  live: "live",
  building: "building",
  planned: "planned",
  paused: "paused",
} as const;

export interface CreateEcosystemProjectBody {
  name: string;
  category: CreateEcosystemProjectBodyCategory;
  status: CreateEcosystemProjectBodyStatus;
  description?: string;
  website?: string;
  logoUrl?: string;
  chainIds?: number[];
  tvl?: string;
  users?: number;
}

export type AgentType = (typeof AgentType)[keyof typeof AgentType];

export const AgentType = {
  nexus: "nexus",
  bridge: "bridge",
  oracle: "oracle",
  guardian: "guardian",
  scout: "scout",
  executor: "executor",
} as const;

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  active: "active",
  idle: "idle",
  paused: "paused",
  error: "error",
} as const;

export interface Agent {
  id: number;
  name: string;
  type: AgentType;
  status: AgentStatus;
  description?: string;
  chainIds?: number[];
  transactionsProcessed?: number;
  uptime?: number;
  lastActive?: string;
  createdAt: string;
}

export type CreateAgentBodyType =
  (typeof CreateAgentBodyType)[keyof typeof CreateAgentBodyType];

export const CreateAgentBodyType = {
  nexus: "nexus",
  bridge: "bridge",
  oracle: "oracle",
  guardian: "guardian",
  scout: "scout",
  executor: "executor",
} as const;

export type CreateAgentBodyStatus =
  (typeof CreateAgentBodyStatus)[keyof typeof CreateAgentBodyStatus];

export const CreateAgentBodyStatus = {
  active: "active",
  idle: "idle",
  paused: "paused",
  error: "error",
} as const;

export interface CreateAgentBody {
  name: string;
  type: CreateAgentBodyType;
  status: CreateAgentBodyStatus;
  description?: string;
  chainIds?: number[];
}

export interface StatsSummary {
  totalAgents: number;
  activeAgents: number;
  totalChains: number;
  totalPartnerships: number;
  activePartnerships: number;
  totalEcosystemProjects: number;
  liveProjects: number;
  totalTransactions: number;
  totalTvl: string;
  networkHealth: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  agent_deployed: "agent_deployed",
  chain_added: "chain_added",
  partnership_signed: "partnership_signed",
  project_launched: "project_launched",
  transaction: "transaction",
  alert: "alert",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description?: string;
  timestamp: string;
  entityId?: number;
  entityType?: string;
}

export interface ChainBreakdown {
  chainName: string;
  chainSymbol: string;
  agentCount: number;
  projectCount: number;
  partnershipCount: number;
  tvl?: string;
}