# lib/api-zod/src/generated/types/chainBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Web3 Global Integration API
 * OpenAPI spec version: 0.1.0
 */

export interface ChainBreakdown {
  chainName: string;
  chainSymbol: string;
  agentCount: number;
  projectCount: number;
  partnershipCount: number;
  tvl?: string;
}