# lib/api-zod/src/generated/types/createAgentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Web3 Global Integration API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAgentBodyStatus } from "./createAgentBodyStatus";
import type { CreateAgentBodyType } from "./createAgentBodyType";

export interface CreateAgentBody {
  name: string;
  type: CreateAgentBodyType;
  status: CreateAgentBodyStatus;
  description?: string;
  chainIds?: number[];
}