# lib/api-zod/src/generated/types/createChainBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Web3 Global Integration API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateChainBodyStatus } from "./createChainBodyStatus";

export interface CreateChainBody {
  name: string;
  symbol: string;
  chainId?: number;
  rpcUrl?: string;
  status: CreateChainBodyStatus;
  tvl?: string;
}