# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Web3 Global Integration API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalAgents: number;
  activeAgents: number;
  totalChains: number;
  totalPartnerships: number;
  activePartnerships: number;
  totalEcosystemProjects: number;
  liveProjects: number;
  totalTransactions: number;
  totalTvl: string;
  networkHealth: number;
}